pyproject.toml
[tool.poetry]
name = "better-self-api"
version = "0.1.0"
description = ""
authors = ["LJlkdskdjflsa <liliangjya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "1.11.3"
annotated-types = "0.5.0"
anyio = "3.7.1"
bcrypt = "4.0.1"
click = "8.1.7"
colorama = "0.4.6"
ecdsa = "0.18.0"
fastapi = "0.103.0"
greenlet = "2.0.2"
h11 = "0.14.0"
idna = "3.4"
mako = "1.2.4"
markupsafe = "2.1.3"
passlib = "1.7.4"
psycopg2 = "2.9.7"
pyasn1 = "0.5.0"
pydantic = "^2"
pydantic-settings = "2.0.3"
pydantic-core = "^2.18.4"
python-dotenv = "1.0.0"
python-jose = "3.3.0"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.20"
starlette = "0.27.0"
typing-extensions = "^4.12.2"
uvicorn = "0.23.2"
line-bot-sdk = "^3.11.0"
langchain = "^0.2.5"
tenacity = "8.3.0"
fastapi-profiler = "^1.4.1"
stripe = "^9.12.0"
firebase-admin = "^6.5.0"
structlog = "^24.2.0"
langchain-anthropic = "0.1.12"
anthropic = "0.23.0"
llama-parse = "^0.4.5"
pypdf = "^4.2.0"
cuid2 = "^2.0.1"
python-multipart = "^0.0.9"
langchain-community = "^0.2.6"
langchainhub = "^0.1.20"


[tool.poetry.group.dev.dependencies]
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
